add commonpage & update webapi
