+++ hello-movie/src/pages/CommonPage/index.js
@@ -1,0 +1,1 @@
+export { default } from "./CommonPage";
